organize all file in event
