+++ src/main/java/dev/it22/kmitl/reg/ui/Homepage_regis/AdminAddSubject.java
@@ -1,5 +1,4 @@
 package dev.it22.kmitl.reg.ui.Homepage_regis;
-import dev.it22.kmitl.reg.ui.event.AdminAddEvent;
 import dev.it22.kmitl.reg.utils.Config;
 import dev.it22.kmitl.reg.utils.CustomCombobox;
 import dev.it22.kmitl.reg.utils.RoundedButton;
