+++ src/main/java/dev/it22/kmitl/reg/ui/event/admin/AdminAddEvent.java
@@ -1,5 +1,6 @@
-package dev.it22.kmitl.reg.ui.event;
+package dev.it22.kmitl.reg.ui.event.admin;
 
+import dev.it22.kmitl.reg.ui.event.calendar.EventPage;
 import dev.it22.kmitl.reg.utils.*;
 
 import java.awt.*;
@@ -118,7 +119,7 @@ public class AdminAddEvent extends EventPage implements ActionListener {
             frame.getContentPane().removeAll();
             frame.revalidate();
             frame.repaint();
-            new AdminCalendarPage (frame);
+            new AdminCalendarPage(frame);
             if (ev.getSource() == yes) {
                 dialog.setVisible(false);
             }
