+++ src/main/java/dev/it22/kmitl/reg/ui/event/admin/AdminCalendarPage.java
@@ -1,6 +1,9 @@
-package dev.it22.kmitl.reg.ui.event;
+package dev.it22.kmitl.reg.ui.event.admin;
 
 import com.formdev.flatlaf.FlatLightLaf;
+import dev.it22.kmitl.reg.ui.event.calendar.calendarTable;
+import dev.it22.kmitl.reg.ui.event.component.newHeader;
+import dev.it22.kmitl.reg.ui.event.calendar.eventCategory;
 import dev.it22.kmitl.reg.utils.Config;
 
 import javax.swing.*;
