+++ src/main/java/dev/it22/kmitl/reg/ui/event/calendar/CalendarPage.java
@@ -1,6 +1,7 @@
-package dev.it22.kmitl.reg.ui.event;
+package dev.it22.kmitl.reg.ui.event.calendar;
 
 import com.formdev.flatlaf.FlatLightLaf;
+import dev.it22.kmitl.reg.ui.event.component.newHeader;
 import dev.it22.kmitl.reg.utils.Config;
 
 import javax.swing.*;
