+++ src/main/java/dev/it22/kmitl/reg/ui/event/calendar/EventPage.java
@@ -1,4 +1,4 @@
-package dev.it22.kmitl.reg.ui.event;
+package dev.it22.kmitl.reg.ui.event.calendar;
 
 import dev.it22.kmitl.reg.utils.Config;
 import dev.it22.kmitl.reg.utils.CustomCombobox;
@@ -7,7 +7,6 @@ import dev.it22.kmitl.reg.utils.RoundedTextField;
 
 import javax.swing.*;
 import java.awt.*;
-import java.awt.event.ActionListener;
 import java.awt.event.FocusEvent;
 import java.awt.event.FocusListener;
