+++ src/main/java/dev/it22/kmitl/reg/ui/event/calendar/calendarTable.java
@@ -1,5 +1,6 @@
-package dev.it22.kmitl.reg.ui.event;
+package dev.it22.kmitl.reg.ui.event.calendar;
 
+import dev.it22.kmitl.reg.ui.event.admin.EditEventPage;
 import dev.it22.kmitl.reg.utils.Config;
 
 import javax.swing.*;
