+++ src/main/java/dev/it22/kmitl/reg/ui/event/component/changeColorCheckbox.java
@@ -1,6 +1,5 @@
-package dev.it22.kmitl.reg.ui.event;
+package dev.it22.kmitl.reg.ui.event.component;
 
-import dev.it22.kmitl.reg.utils.Config;
 import javax.swing.*;
 import java.awt.*;
