+++ src/main/java/dev/it22/kmitl/reg/ui/event/component/headerMenu.java
@@ -1,4 +1,4 @@
-package dev.it22.kmitl.reg.ui.event;
+package dev.it22.kmitl.reg.ui.event.component;
 
 import dev.it22.kmitl.reg.utils.Config;
